resolve circular dependencies in actions
