+++ packages/ketcher-core/src/application/editor/actions/erase.ts
@@ -33,10 +33,8 @@ import { removeAtomFromSgroupIfNeeded, removeSgroupIfNeeded } from './sgroup';
 import { Action } from './action';
 import assert from 'assert';
 import { atomGetDegree, formatSelection } from './utils';
-import {
-  fromBondStereoUpdate,
-  removeAttachmentPointFromSuperatom,
-} from '../actions/bond';
+import { removeAttachmentPointFromSuperatom } from '../actions/bond';
+import { fromBondStereoUpdate } from './bondStereo';
 import { fromFragmentSplit } from './fragment';
 import { fromRGroupAttachmentPointDeletion } from './rgroupAttachmentPoint';
 import { ReStruct } from 'application/render';
