+++ packages/ketcher-core/src/application/editor/actions/index.ts
@@ -1,8 +1,10 @@
 export * from './action';
 export * from './aromaticFusing';
 export * from './atom';
+export * from './atomMerge';
 export * from './basic';
 export * from './bond';
+export * from './bondStereo';
 export * from './chain';
 export * from './closelyFusing';
 export * from './erase';
