+++ packages/ketcher-core/src/application/editor/actions/utils.ts
@@ -27,7 +27,7 @@ import { Vec2 } from 'domain/entities/vec2';
 import closest from '../shared/closest';
 import { ReStruct } from 'application/render';
 import { selectionKeys } from '../shared/constants';
-import { EditorSelection } from '../editor.types';
+import type { EditorSelection } from '../editor.types';
 export type AtomType = 'single' | 'list' | 'pseudo';
 export type AtomAttributeName = keyof AtomAttributes;
 export type AtomQueryPropertiesName = keyof AtomQueryProperties;
